+++ src/Storage/Storage.Management/ChangeLog.md
@@ -30,6 +30,14 @@
     - `New-AzStorageShareSasToken`
     - `New-AzStorageQueueSasToken`
     - `New-AzStorageTableSasToken`
+* Migrated following Azure Queue dataplane cmdlets from 'Microsoft.Azure.Storage.Queue 11.2.2' to 'Azure.Storage.Queues 12.16.0'
+    - `New-AzStorageQueue`
+    - `Get-AzStorageQueue`
+    - `Remove-AzStorageQueue`
+    - `New-AzStorageQueueStoredAccessPolicy`
+    - `Get-AzStorageQueueStoredAccessPolicy`
+    - `Set-AzStorageQueueStoredAccessPolicy`
+    - `Remove-AzStorageQueueStoredAccessPolicy`
 
 ## Version 5.10.1
 * Added warning messages for an upcoming breaking change that the output Permissions will be changed to a string when creating and updating a Queue access policy